+++ lib/ui/home/HomeUpcomingEventsWidget.dart
@@ -362,7 +362,7 @@ class _HomeUpcomingEventsWidgetState extends State<HomeUpcomingEventsWidget> imp
 
   void _navigateToExploreEvents() {
     Analytics().logSelect(target: "HomeUpcomingEvents View all events");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => ExplorePanel(initialTab: ExploreTab.Events)));
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => ExplorePanel(initialItem: ExploreItem.Events)));
   }
 }